Remove SQLAlchemy-based database management modules and refactor with custom database connection managers.
