+++ src/lifespan.py
@@ -3,15 +3,14 @@ from typing import AsyncIterator
 
 from fastapi import FastAPI
 
-from src.database.connection_factory import create_connection
-from src.database.models import CommonMetadata
+from src.database.db_connector import create_connection
 
 
 @asynccontextmanager
 async def lifespan_routine(app: FastAPI) -> AsyncIterator[None]:
     # put any other startup init here (DB pools, caches, etc.)
     db_connection = create_connection()
-    db_connection.create_all(CommonMetadata)
+    db_connection.test_db_connection()
 
     try:
         yield
